+++ src/webapp/src/components/pipelines/ExtractorSettingsModal.vue
@@ -25,9 +25,6 @@ export default {
         this.extractorInFocusConfiguration.settings.length === 0
       )
     },
-    extractorLacksConfigSettingsAndIsInstalled() {
-      return !this.isInstalling && this.extractorLacksConfigSettings
-    },
     extractor() {
       return this.getInstalledPlugin('extractors', this.extractorNameFromRoute)
     },
@@ -69,10 +66,10 @@ export default {
         }
         this.addPlugin(config).then(() => {
           this.prepareExtractorConfiguration()
-          this.installPlugin(config)
+          this.installPlugin(config).then(this.checkAutoAdvance)
         })
       } else {
-        this.prepareExtractorConfiguration()
+        this.prepareExtractorConfiguration().then(this.checkAutoAdvance)
       }
     })
   },
@@ -88,8 +85,13 @@ export default {
         this.$router.push({ name: 'extractors' })
       }
     },
+    checkAutoAdvance() {
+      if (this.extractorLacksConfigSettings) {
+        this.saveConfigAndBeginEntitySelection()
+      }
+    },
     prepareExtractorConfiguration() {
-      this.$store.dispatch(
+      return this.$store.dispatch(
         'configuration/getExtractorConfiguration',
         this.extractorNameFromRoute
       )
@@ -161,28 +163,10 @@ export default {
             </p>
           </div>
         </template>
-
-        <template v-if="extractorLacksConfigSettingsAndIsInstalled">
-          <div class="content">
-            <p>{{ extractorNameFromRoute }} doesn't require configuration.</p>
-            <ul>
-              <li>Click "Next" to advance</li>
-              <li>Click "Cancel" to install other extractors</li>
-            </ul>
-          </div>
-        </template>
       </section>
       <footer class="modal-card-foot buttons is-right">
         <button class="button" @click="close">Cancel</button>
         <button
-          v-if="extractorLacksConfigSettingsAndIsInstalled"
-          class="button is-interactive-primary"
-          @click="saveConfigAndBeginEntitySelection"
-        >
-          Next
-        </button>
-        <button
-          v-else
           class="button is-interactive-primary"
           :disabled="!isSaveable"
           @click="saveConfigAndBeginEntitySelection"